pyproject.toml
[project]
name = "WebLogTriage"
version = "1.0.1"
description = "A tool to triage web server logs for suspicious activity."
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.7"
dependencies = [
    "pandas",
    "numpy",
    "pytz",
    "PyYAML",
    "six",
    "joblib",
    "prettytable",
    "termcolor"
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
]
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
